+++ e2e-network/expect-invoke-rest.sh
@@ -1,6 +1,9 @@
 #!/usr/bin/env bash
 
-rest_api_url="$1"
+set -eu
+
+rest_api_url="$(echo "$1" | awk '{print $1}')"
+access_token="$(echo "$1" | awk '{print $2}')"
 channel="$2"
 chaincode="$3"
 method="$4"
@@ -18,18 +21,14 @@ label="Invoke $rest_api_url/invoke/$channel/$chaincode $method"
 echo ""
 echo "➜ testing: $label"
 
-enroll_admin_response="$(
-  curl \
-    -s \
-    --request POST \
-    --url "$rest_api_url/user/enroll" \
-    --header 'Content-Type: application/json' \
-    --data '{"id": "admin", "secret": "adminpw"}'
-)"
-
-echo "$enroll_admin_response"
-
-access_token="$(echo "$enroll_admin_response" | jq -r '.token')"
+if [ -z "$access_token" ]; then
+  the_same_dir="$(cd "$(dirname "$0")" && pwd)"
+  enroll_admin_response="$("$the_same_dir/expect-ca-rest.sh" "$rest_api_url/user/enroll" '' '{"id": "admin", "secret": "adminpw"}' "token")"
+  echo "enroll admin response: $enroll_admin_response"
+  access_token="$(echo "$enroll_admin_response" | jq -r '.token')"
+else
+  echo "using provided token: $access_token"
+fi
 
 response=$(
   curl \