+++ fablo.sh
@@ -28,44 +28,59 @@ getDefaultFabloConfig() {
   fi
 }
 
+getSnapshotPath() {
+  path="${1:-'snapshot'}"
+  if echo "$path" | grep -q "tar.gz$"; then
+    echo "$path"
+  else
+    echo "$path.fablo.tar.gz"
+  fi
+}
+
 printHelp() {
   echo "Fablo -- kick-off and manage your Hyperledger Fabric network
 
 Usage:
-  fablo.sh init [node] [rest]
+  fablo init [node] [rest]
     Creates simple Fablo config in current directory with optional Node.js sample chaincode and REST API.
 
-  fablo.sh generate [/path/to/fablo-config.json|yaml [/path/to/fablo/target]]
+  fablo generate [/path/to/fablo-config.json|yaml [/path/to/fablo/target]]
     Generates network configuration files in the given directory. Default config file path is '\$(pwd)/fablo-config.json' or '\$(pwd)/fablo-config.yaml', default (and recommended) directory '\$(pwd)/fablo-target'.
 
-  fablo.sh up [/path/to/fablo-config.json|yaml]
+  fablo up [/path/to/fablo-config.json|yaml]
     Starts the Hyperledger Fabric network for given Fablo configuration file, creates channels, installs and instantiates chaincodes. If there is no configuration, it will call 'generate' command for given config file.
 
-  fablo.sh <down | start | stop>
+  fablo <down | start | stop>
     Downs, starts or stops the Hyperledger Fabric network for configuration in the current directory. This is similar to down, start and stop commands for Docker Compose.
 
-  fablo.sh reboot
+  fablo reboot
     Downs and ups the network. Network state is lost, but the configuration is kept intact.
 
-  fablo.sh prune
+  fablo prune
     Downs the network and removes all generated files.
 
-  fablo.sh recreate [/path/to/fablo-config.json|yaml]
+  fablo recreate [/path/to/fablo-config.json|yaml]
     Prunes and ups the network. Default config file path is '\$(pwd)/fablo-config.json' or '\$(pwd)/fablo-config.yaml'.
 
-  fablo.sh chaincode upgrade <chaincode-name> <version>
+  fablo chaincode upgrade <chaincode-name> <version>
     Upgrades and instantiates chaincode on all relevant peers. Chaincode directory is specified in Fablo config file.
 
-  fablo.sh channel --help
+  fablo channel --help
     To list available channel query options which can be executed on running network.
 
-  fablo.sh use [version]
+  fablo snapshot <target-snapshot-path>
+    Creates a snapshot of the network in target path. The snapshot contains all network state, including transactions and identities.
+
+  fablo restore <source-snapshot-path>
+    Restores the network from a snapshot.
+
+  fablo use [version]
     Updates this Fablo script to specified version. Prints all versions if no version parameter is provided.
 
-  fablo.sh <help | --help>
+  fablo <help | --help>
     Prints the manual.
 
-  fablo.sh version [--verbose | -v]
+  fablo version [--verbose | -v]
     Prints current Fablo version, with optional details."
 }
 
@@ -159,6 +174,45 @@ networkUp() {
   "$FABLO_TARGET/fabric-docker.sh" up
 }
 
+executeFabloDockerCommand() {
+  if [ ! -d "$FABLO_TARGET" ]; then
+    echo "Error: This command needs the network to be generated at '$FABLO_TARGET'! Execute 'generate' or 'up' command."
+    exit 1
+  fi
+
+  echo "Executing Fablo docker command: $1"
+  "$FABLO_TARGET/fabric-docker.sh" "$1" "$2" "$3" "$4" "$5" "$6" "$7" "$8"
+}
+
+createSnapshot() {
+  archive="$(getSnapshotPath "$1")"
+  echo "Creating network snapshot in '$archive'"
+
+  if [ -f "$archive" ]; then
+    echo "Error: Snapshot file '$archive' already exists!"
+    exit 1
+  fi
+
+  executeFabloDockerCommand snapshot "$FABLO_TEMP_DIR"
+  (cd "$FABLO_TEMP_DIR" && tar czf tmp.tar.gz *)
+  mv "$FABLO_TEMP_DIR/tmp.tar.gz" "$archive"
+  echo "📦 Created snapshot at '$archive'!"
+}
+
+restoreSnapshot() {
+  archive="$(getSnapshotPath "$1")"
+  echo "📦 Restoring network from '$archive'"
+
+  if [ ! -f "$archive" ]; then
+    echo "Fablo snapshot file '$archive' does not exist!"
+    exit 1
+  fi
+
+  tar -xf "$archive" -C "$FABLO_TEMP_DIR"
+  "$FABLO_TEMP_DIR/fablo-target/fabric-docker.sh" clone-to "$COMMAND_CALL_ROOT"
+  echo "📦 Network restored from '$archive'! Execute 'start' command to run it."
+}
+
 if [ -z "$COMMAND" ]; then
   printHelp
   exit 1
@@ -194,7 +248,12 @@ elif [ "$COMMAND" = "recreate" ]; then
   networkPrune
   networkUp "$2"
 
+elif [ "$COMMAND" = "snapshot" ]; then
+  createSnapshot "$2"
+
+elif [ "$COMMAND" = "restore" ]; then
+  restoreSnapshot "$2"
+
 else
-  echo "Executing Fablo docker command: $COMMAND"
-  "$FABLO_TARGET/fabric-docker.sh" "$COMMAND" "$2" "$3" "$4" "$5" "$6" "$7" "$8"
+  executeFabloDockerCommand "$COMMAND" "$2" "$3" "$4" "$5" "$6" "$7" "$8"
 fi