+++ src/setup-docker/templates/fabric-docker/snapshot-scripts.sh
@@ -1,0 +1,81 @@
+#!/usr/bin/env bash
+
+function __getSnapshotNodes() {
+  network_name="${COMPOSE_PROJECT_NAME}_basic"
+  docker ps --format "{{.Names}}" --filter "network=$network_name" --all |
+    grep -v "cli\." |
+    grep -v "dev-peer" |
+    grep -v "ca\." |
+    grep -v "couchdb\." |
+    grep -v "fablo-rest"
+}
+
+function __getCANodes() {
+  network_name="${COMPOSE_PROJECT_NAME}_basic"
+  docker ps --format "{{.Names}}" --filter "network=$network_name" --all | grep "ca\."
+}
+
+function __createSnapshot() {
+  cd "$FABLO_NETWORK_ROOT/.."
+  backup_dir="${1:-"snapshot-$(date -u +"%Y%m%d%H%M%S")"}"
+
+  if [ -d "$backup_dir" ] && [ "$(ls -A "$backup_dir")" ]; then
+    echo "Error: Directory '$backup_dir' already exists and is not empty!"
+    exit 1
+  fi
+
+  mkdir -p "$backup_dir"
+  cp -R ./fablo-target "$backup_dir/"
+
+  for node in $(__getCANodes); do
+    echo "Saving state of $node..."
+    mkdir -p "$backup_dir/$node"
+    docker cp "$node:/etc/hyperledger/fabric-ca-server/fabric-ca-server.db" "$backup_dir/$node/fabric-ca-server.db"
+  done
+
+  for node in $(__getSnapshotNodes); do
+    echo "Saving state of $node..."
+    docker cp "$node:/var/hyperledger/production/" "$backup_dir/$node/"
+  done
+}
+
+function __cloneSnapshot() {
+  cd "$FABLO_NETWORK_ROOT/.."
+  target_dir="$1"
+
+  if [ -d "$target_dir/fablo-target" ]; then
+    echo "Error: Directory '$target_dir/fablo-target' already exists! Execute 'fablo prune' to remove the current network."
+    exit 1
+  fi
+
+  cp -R ./fablo-target "$target_dir/fablo-target"
+  (cd "$target_dir/fablo-target/fabric-docker" && docker-compose up --no-start)
+
+  network_name="${COMPOSE_PROJECT_NAME}_basic"
+
+  for node in $(__getCANodes); do
+    echo "Restoring $node..."
+    if [ ! -d "$node" ]; then
+      echo "Warning: Cannot restore '$node', directory does not exist!"
+    else
+      docker cp "./$node/fabric-ca-server.db" "$node:/etc/hyperledger/fabric-ca-server/fabric-ca-server.db"
+    fi
+  done
+
+  for node in $(__getSnapshotNodes); do
+    echo "Restoring $node..."
+    if [ ! -d "$node" ]; then
+      echo "Warning: Cannot restore '$node', directory does not exist!"
+    else
+      docker cp "./$node/" "$node:/var/hyperledger/production/"
+    fi
+  done
+}
+
+createSnapshot() {
+  (set -eu && __createSnapshot "$1")
+}
+
+cloneSnapshot() {
+  (set -eu && __cloneSnapshot "$1")
+}